+++ app/controllers/api/ui/v1/authentication_controller.rb
@@ -1,16 +1,16 @@
 # https://github.com/anusharanganathan/data2paper/wiki/Dual-authentication-using-Devise-and-JWT
-class API::Ui::V1::AuthenticationController < API::Ui::BaseController
-  skip_before_action :authenticate_request, only: [:authenticate]
+class Api::Ui::V1::AuthenticationController < Api::Ui::BaseController
+  skip_before_action :authenticate_request, only: [:authenticate], raise: false
 
   def authenticate
-    require_relative Rails.root.join 'app/services/user_authentication/authenticate_user.rb'
-    service = AuthenticateUser.call(params[:email],
-                                    params[:password],
-                                    request)
+    require_relative Rails.root.join 'app/services/user_authentication/authenticate_api_user.rb'
+    service = AuthenticateApiUser.call(params[:email],
+                                       params[:password],
+                                       request)
     if service.success?
       @auth_token = service.result
       @user = service.user
-      if @user.confirmed_at.nil?
+      if @user.confirmed_at.nil?y
         render json: { error: t('ui.users.alerts.unconfirmed') }, status: :unauthorized
       else
         @user.log_devise_action('login success')