+++ app/controllers/api/ui/v1/registrations_controller.rb
@@ -1,4 +1,4 @@
-class API::Ui::V1::RegistrationsController < Devise::RegistrationsController
+class Api::Ui::V1::RegistrationsController < Devise::RegistrationsController
   include JWTAuthenticable
   include GuestControllable
   before_action :check_and_sanitize_email_confirmation, only: [:create]
@@ -9,21 +9,19 @@ class API::Ui::V1::RegistrationsController < Devise::RegistrationsController
   respond_to :json
 
   def create
-    customer = Customer.new(customer_params.except(:email_confirmation))
-    customer.organization_id = current_organization.id
-    customer.origin = Customer::ORIGIN_WEB
+    user = User.new(user_params.except(:email_confirmation))
     result = begin
-      Customer.transaction do
-        if customer.save
-          require_relative Rails.root.join 'app/services/guest_authentication/authenticate_customer.rb'
-          service = AuthenticateUser.call(customer_params[:email],
-                                          customer_params[:password],
-                                          customer.organization_id)
+      User.transaction do
+        if user.save
+          require_relative Rails.root.join 'app/services/user_authentication/authenticate_api_user.rb'
+          service = AuthenticateApiUser.call(user_params[:email],
+                                             user_params[:password],
+                                             request)
           @auth_token = service.result
-          @customer = service.customer
+          @user = service.user
         end
         # Oltre a crearlo, dobbiamo verificare che si possa loggare
-        raise 'Customer not authenticated' unless @customer
+        raise 'User not authenticated' unless @user
         true
       end
     rescue
@@ -33,7 +31,7 @@ class API::Ui::V1::RegistrationsController < Devise::RegistrationsController
       render :create, status: 201
     else
       warden.custom_failure!
-      render json: { errors: customer.errors }, status: 422
+      render json: { errors: user.errors }, status: 422
     end
   end
 
@@ -41,32 +39,26 @@ class API::Ui::V1::RegistrationsController < Devise::RegistrationsController
 
   private
 
-  def customer_params
-    if params[:customer].present?
-      params.require(:customer).permit(:email, :email_confirmation, :password, :password_confirmation,
-                                       :fiscal_code, :first_name, :last_name, :business_name,
-                                       :document_number, :customer_type, :language, :category, :phone_prefix, :phone,
-                                       :address, :cap,
-                                       :birthdate, :sex, :identification_document_id, :document_issued_on,
-                                       :document_expiring_date, :place_of_birth_id, :country_of_birth_id,
-                                       :document_issued_location_id, :citizenship_id, :country_of_residence_id,
-                                       :city_of_residence_id, :ok_marketing, :pending_order_uuid)
-    else
-      {}
-    end
+  def user_params
+    params.require(:user).permit(:email, :email_confirmation, :password, :password_confirmation,
+                                 :firstname, :lastname, :phone_prefix, :phone, :birthdate, :gender, :format )
+    # if params[:user].present?
+    # else
+    #   {}
+    # end
   end
 
   def check_and_sanitize_email_confirmation
     errors = {}
-    if customer_params[:email].blank?
+    if user_params[:email].blank?
       errors[:email] ||= []
       errors[:email] << 'is blank'
     end
-    if customer_params[:email_confirmation].blank?
+    if user_params[:email_confirmation].blank?
       errors[:email_confirmation] ||= []
       errors[:email_confirmation] << 'is blank'
     end
-    if errors.blank? && customer_params[:email] != customer_params[:email_confirmation]
+    if errors.blank? && user_params[:email] != user_params[:email_confirmation]
       errors[:email] ||= []
       errors[:email] << 'does not coincide with email_confirmation'
     end