+++ app/controllers/users_controller.rb
@@ -1,6 +1,6 @@
 class UsersController < ApplicationController
   helper VueHelper
-  # skip_before_action :basic_authenticate
+
   layout 'users'
 
   def index